timeout screen

+++ Assets/Scripts/JsonClasses.cs
@@ -26,7 +26,7 @@ public class StepResponse : RequestResponse
     public float[] observation;
     public bool done;
     public int fps;
-    public int duration_pause;
+    public float duration_pause;
     public float distance_from_goal;
     public int human_action;
     public int agent_action;
